skills/greenlight-testing/requirements.txt
# Greenlight Skill - Python Dependencies
# Install with: pip install -r requirements.txt

pyyaml>=6.0          # YAML parsing for test specs
